Introduce full list of research guides as HTML
This commit introduces the full list of 365 research guides and updates
the approach to progressively enhance these to a single-page app.

Also removes some filtering by recommended since this status is not a
property of the guide itself but rather a property of the relationship
between a guide and a subject. Needs a rethink, so removed for clarity.